requirements.txt
flask
Pillow
opennsfw2
requests
opencv-python-headless
gunicorn